Implement Valid-Until for security pocket
Implement the Valid-Until field for all security pockets;
including existing ones.

Set the Valid-Until field two weeks after the publication
Date field, and republish the Release file if it will expire
in less than a week.

A similar change probably needs to be made for ESM publishing,
I haven't checked that. This works for the main archive at the
very least.

LP: #716535